gogo: Move to own pkg
