+++ internal/serializers/gogo/structdef-gogo.proto
@@ -2,7 +2,7 @@ syntax = "proto2";
 
 package goserbench;
 
-option go_package = "github.com/alecthomas/go_serialization_benchmarks;goserbench";
+option go_package = "github.com/alecthomas/go_serialization_benchmarks/internal/gogo;gogo";
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
